# New Bloom filter implementation for full and partitioned filters