--- a/include/rocksdb/table.h
+++ b/include/rocksdb/table.h
@@ -272,7 +272,7 @@ struct BlockBasedTableOptions {
   // 5 -- Can be read by RocksDB's versions since X.X.X (something after 6.4.6)
   // Full and partitioned filters use a generally faster and more accurate
   // Bloom filter implementation, with a different schema.
-  uint32_t format_version = 5;  // TODO: back to 2
+  uint32_t format_version = 2;
 
   // Store index blocks on disk in compressed format. Changing this option to
   // false  will avoid the overhead of decompression if index blocks are evicted