--- a/table/block_based/parsed_full_filter_block.h
+++ b/table/block_based/parsed_full_filter_block.h
@@ -25,7 +25,7 @@ class ParsedFullFilterBlock {
     return filter_bits_reader_.get();
   }
 
-  // TODO: consider memory usage of FullFilterBitsReader
+  // TODO: consider memory usage of the FilterBitsReader
   size_t ApproximateMemoryUsage() const {
     return block_contents_.ApproximateMemoryUsage();
   }