--- a/table/plain/plain_table_bloom.h
+++ b/table/plain/plain_table_bloom.h
@@ -4,8 +4,6 @@
 //  (found in the LICENSE.Apache file in the root directory).
 #pragma once
 
-#include <folly/ConstexprMath.h>
-#include <memory>
 #include <string>
 #include <vector>
 
@@ -15,6 +13,9 @@
 #include "util/bloom_impl.h"
 #include "util/hash.h"
 
+#include "third-party/folly/folly/ConstexprMath.h"
+
+#include <memory>
 
 namespace rocksdb {
 class Slice;